Added Reihenbestände function
As the relevant forests consist mainly of a single tree type, a main tree type is now selected for every cluster or dense forest. Selection of different tree types can be set to sepcific likelihoods at the top of synth_forest.py, while a second likelihood p in the place_tree function handles if the main tree is actually used or another tree type is selected at random.