Clusters now use single tree, smoothing not between trees
When using the function fill_with_trees inside a cluster, fill_with_a_single_tree can be enabled. This uses only a single tree as the main tree (instead of a tree type) instead of a tree type with different trees. Further, edge_smoothing is now only performed on edges of trees not overlapping with other trees. This is achieved by buffering the edge_mask and then masking it with the tree mask, leaving only the buffer around the trees intersecting with background.